requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiogram-dialog==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.7.1
yarl==1.9.2

asyncpg~=0.29.0
alembic~=1.13.1